Basic Complete

+++ ExpenseRecordMVC/ExpenseService/CategoryService.cs
@@ -7,10 +7,12 @@ namespace ExpenseService
     public class CategoryService : ICategoryService
     {
         private readonly ICategoryRepository categoryRepository;
+        private readonly IExpenseRepository expenseRepository;
 
-        public CategoryService(ICategoryRepository categoryRepository)
+        public CategoryService(ICategoryRepository categoryRepository, IExpenseRepository expenseRepository)
         {
             this.categoryRepository = categoryRepository;
+            this.expenseRepository = expenseRepository;
         }
 
         public async Task<List<UserChoice>> GetUserChoicesAsync(string? uname)
@@ -36,7 +38,19 @@ namespace ExpenseService
 
         public async Task<bool?> DeleteChoiceAsync(int id)
         {
-            return await categoryRepository.DeleteChoiceAsync(id);
+            if (await expenseRepository.IsUsedCategory(id))
+            {
+                return false;
+            }
+            else
+            {
+                return await categoryRepository.DeleteChoiceAsync(id);
+            }
+        }
+
+        public async Task<int> GerUserChoiceIdAsync(string? userName, string? CategoryName)
+        {
+            return await categoryRepository.GerUserChoiceIdAsync(userName, CategoryName);
         }
     }
 }
